requirements.txt
mss
opencv-python
numpy
pyautogui
pillow
pygetwindow
Quartz